binary_sensor/rest: do not rewrite rest data while parsing
Currently parsing the REST data rewrites the values themselves
while applying the templates. Because of this, if `is_on` is
called twice, the second time it will always return false,
as the template will no longer apply.
Instead, manipulate a temporary variable to keep the REST response
immutable. This is the same sort of behaviour `update()` in
`sensor/rest.py`.

Signed-off-by: Gergely Imreh <imrehg@gmail.com>